commit fix
removed prebuilt binary and added failsafe for sript in case binary is missing

+++ custommapping.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
+# check if binary has been built
+if [ ! -f $SCRIPTPATH/src/wacom-tool ]; then
+    echo "[!] You must first build the binary in the src directory"
+    exit 1
+fi
+
+# load settings
 SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
 source $SCRIPTPATH/settings
 
+# warning messages
 echo "[!] Do NOT resize the GUI, it is supposed to have the same ratio as your screen!"
 echo "[.] Press <Enter> to save the current mapping"
 echo "[.] Press 'r' to adjust the mapped area to the ratio of your tablet"
@@ -10,4 +18,4 @@ echo ""
 read -n 1 -s -r -p "Press Any Key to Continue.."
 echo ""
 
-$SCRIPTPATH/data/wacom-tool $SCRIPTPATH
+$SCRIPTPATH/src/wacom-tool $SCRIPTPATH
